Refactor phimm_utils
